+++ perf.f90
@@ -1,16 +1,16 @@
 ! https://github.com/JuliaLang/julia/blob/master/test/perf/micro/perf.f90
 module perf
-    use, intrinsic :: iso_fortran_env, only : dp=>REAL64,i64=>INT64, stderr=>error_unit
+    use, intrinsic :: iso_fortran_env, only : REAL64,INT64, stderr=>error_unit
     implicit none
 contains
 
-    real(dp) function sysclock2ms(t)
+    real(real64) function sysclock2ms(t)
     ! Convert a number of clock ticks, as returned by system_clock called
-    ! with integer(i64) arguments, to milliseconds
+    ! with integer(int64) arguments, to milliseconds
 
-        integer(i64), intent(in) :: t
-        integer(i64) :: rate
-        real(dp) ::  r
+        integer(int64), intent(in) :: t
+        integer(int64) :: rate
+        real(real64) ::  r
         call system_clock(count_rate=rate)
         r = 1000.d0 / rate
         sysclock2ms = t * r
